--- apps/vps/src/routes/admin/admin.index.ts
@@ -1,10 +1,0 @@
-import { createRouter } from '@/lib/create-app'
-import * as handlers from './admin.handlers'
-import * as routes from './admin.routes'
-
-const router = createRouter()
-  .openapi(routes.getAdminOverview, handlers.getAdminOverview)
-  .openapi(routes.simulateFrontendError, handlers.simulateFrontendError)
-  .openapi(routes.getNewsletterSubscribers, handlers.getNewsletterSubscribers)
-
-export default router
