+++ packages/api/package.json
@@ -4,6 +4,7 @@
   "type": "module",
   "exports": {
     "./api": "./src/api.ts",
+    "./admin": "./src/admin.ts",
     "./health": "./src/health.ts",
     "./music": "./src/music.ts",
     "./search": "./src/search.ts",
